mdnsd: Do not change default TTL for mDNS traffic, link-local only
mDNS, and everything else in 224.0.0.x, is supposed to be link-local
only and not be routable.  Operators that want to route mDNS traffic
should rethink and try to bridge, or use a layer-2 tunnel, instead.

In Linux it is also possible to mangle the frames with iptables:

    iptables -t mangle -A PREROUTING -i eth0 -d 224.0.0.251 -j TTL --ttl-inc 1

Options on other operating systems likely vary.  Nevertheless, as the
upstream project we should not default to TTL = 255.  We could allow
a pull request for a `-t TTL` command line option though.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>