Increase socket receive buffer size
At startup of a network of many devices there's a lot of initial mDNS
traffic.  Increase the receive buffer (double it) since we may be a low
priority task as well.

On Linux the getsockopt() for SO_RCVBUF returns double the set size, due
to housekeeping in the kernel.  So we don't have to do any adjustments
when setting it.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>