Fix #14: Join 224.0.0.251 on the selected interface
When joining the mDNS multicast group we need to join it on the selected
interface, or we will not recive any multicast if the system does not
have a default route, or a net route to 224.0.0.0/24.

This is the main issue reported in issue #14, so I'm closing it with
this commit.  A more comprehensive (multi-homed) solution is described
in issue #8.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>