Follow-up to bb88467d: continue parsing incoming messages
We cannot let message_parse() or mdnsd_in() return 1 back to main(),
only recvfrom() should do that.  Because if we do, then main() will
exit() the application and cause it to be restarted without reason.

An invalid message does not indicate all other incoming messages are
invalid, so just discard bad messages and continue parsing instead.

Continued problems with recvfrom() may however indicate problems with
the socket, and the interface we're listening on.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>