mdnsd: create PID file when starting up, touch on .conf reload
Traditional UNIX daemons create a PID file in /var/run/$0.pid.  This is
a courtesy to command line applications and scripts.  By practice it is
also a way to monitor and check for stale daemons.

Some init systems rely on the creation of the PID file, and it being
updated after SIGHUP to reload the daemon .conf file.  The creation of
the PID file is a signal to the rest of the system that the daemon is up
and running, ready to service requests, which in turn can be used to
synchronize other system processes.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>