libmdnsd: follow-up to 1dbcbb00, only drop known looped back packets
This fix, albeit a bit costly due to the extra processing of our own
announcements, actually allows us to run mdns-scan and mquery on the
same device.  Without this fix *all* mDNS packet originating from the
same address as ourselves are dropped.

I've tried the pktinfo trick when sending, flipping the multicast LOOP
flag, and lots of other workarounds.  This is the most portable and
least intrusive to get the full functionality of detecting collisions
and allowing local scan requests.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>