Fix #8: add support for running on multiple interfaces
This huge patch adds support for running on multiple interfaces.
There are quite a few changes to the setup and initialization of
the daemon.

Basically, this implmentation is the same as running multiple
instances of mdnsd v0.10, one per interface, sharing the same
.conf files for services to announce.

A later patch may add more granular support for filtering or selecting
which services are to be announced on which interface. Also, currently
mdnsd runs on all interfaces or one.  If you have seven interfaces and
only want to run on two of them, mdnsd won't support you right now.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>