Fix parsing of A record from packet
When parsing an A record from the packet, the `net2long` function is used
to read the IPv4 address and write it into the `in_addr` structure element
`s_addr`. But the definition of the `s_addr` element is to hold the IP
address in network byte order. The `net2long` function is a convenience
function which converts from network byte order to host byte order, and
then moves the buffer pointer forward. But the convenience is too much at
this point, because the value read from the packet should not be converted
but must be kept in network byte order.