libmdnsd: Provide specific methods to add IP addresses to rdata
There was a bit of a mess when adding IPv4 addresses to RR data when
building a message. Which lead to the `message_data_long` function not
being used anymore as it turned unusable.

Instead, revert `message_data_long` to properly adding a long integer
to the message, converting to network byte order as it should. Two new
methods are added to specifically add an IPv4 and an IPv6 address to the
message, taking care of the specifics of the two cases.