mdnsd: Add IPv4 and IPv6 addresses to interfaces and report A and AAAA
An `iface`, representing a system interface, can have an IPv4 and a IPv6
address. This is needed so that the `mdnsd` knows about both address
family types and can answer A and AAAA records for this interface.
When scanning system interfaces run over all addresses to keep adding
IPv4 as well as IPv6 addresses to the same `iface`.

When reading the configuration a AAAA record is added in addition to
the A record. The `mdns` instance gets set the IPv6 address when
setting up the interface.