mdnsd: Only create A records if an IPv4 address is set
When adding IPv6 address support, it can happen that only a IPv4 or
an IPv6 address is set in the future. Then unconditionally adding a
A (or AAAA) record will break, because there will be no RDATA, i.e. IP
address to go along with it. Which will result in malformed packets.

Thus records are only added when a IP is already set. Which is why the
IP now needs to be set on the mdnsd before `conf_init` is called.

This is a bit of a quick and dirty solution. Preferably the function
`mdnsd_set_address` would not only update existing records but also create
a new one if none exists yet. But that means that additional information
like the hostname and the conflict callback need to be provided. I found
no elegant solution for this. Another possibility would be to change its
return value to indicate whether a record needs to be created, because
none was found, and leave it to the caller to create one.

Also adds a test to check if an A record was created. This should also
help with future refactoring of the code to create A/AAAA records.