libmdnsd: Remove A/AAAA records when IP is set to 0
With multiple protocols available it can happen that an active interface
gets all IPv6 or IPv4 addresses removed. In that case the responder should
stop answering with A/AAAA records. Moreover, it should send a goodbye
packet that invalidates the last known IP address.
This is done in `mdnsd_set_ip_address` when the IP address is 0.