libmdnsd: Process messages from/to IPv4 and IPv6 destinations
mDNS messages can be sent over IPv4 or IPv6. `mdnsd_in` is able to store
the source address for both address families, and `mdnsd_out` can fill
in the destination `sockaddr` also for both address families.
When calling `mdnsd_out`, the caller must fill in the address family in
the `to` parameter struct so that the correct data can be chosen and be
filled in to the `to` struct.

This change makes no assumption whether messages of both or only one of
the two address families are processed in the same mdnsd instance. While
it probably make sense to separate the IPv4 and IPv6 domains and serve
them with separate mdnsd instances, this patch allows for messages on
both to be handled by one. This patch does not concern itself with the
question how to properly handle Disco, probing and announcement, when
only one mdnsd instance serves both domains.