test/lib.sh: Test IPv6 connectivity
Test IPv6 connectivity with a ping just as IPv4 connectivity. Because
mdnsd is to be extended to also run over IPv6 this should be tested in
the setup.
Well, almost as IPv4 connectivity. The `ping6` must not use double quotes
around the `${server_addr_ll6}` variable expansion or the address will be
taken as a hostname, which will not resolve to any node.

For the ping to succeed we wait for the double address detection to finish
and the link local IPv6 address become valid, i.e. lose the tentative
state.